// server/models/HousingFiltersApi.tsx
export interface HousingFiltersApi {
    ownerKinds: string[];
    ownerAges: string[];
    multiOwners: string[];
    beneficiaryCounts: string[];
    contactsCounts: string[];
    housingKinds: string[];
    housingStates: string[];
    housingAreas: string[];
    buildingPeriods: string[];
    vacancyDurations: string[];
    isTaxedValues: string[];
    query: string;
}